make api more robust

+++ www/api.php
@@ -23,25 +23,34 @@ try {
     $restInfo = $request->getRestInfo();
     
     $rs = new ResourceServer($storage, $config);
-    $token = $rs->verify($request->getHeader("HTTP_AUTHORIZATION"));
+    if($request->headerExists("HTTP_AUTHORIZATION")) { 
+            
+        $token = $rs->verify($request->getHeader("HTTP_AUTHORIZATION"));
 
-    if($restInfo->match("POST", "authorizations", FALSE)) {
-        $data = json_decode($request->getContent(), TRUE);
-        $storage->addApproval($data['client_id'], $token->resource_owner_id, $data['scope']);
-        $response->setStatusCode(201);
-    } else if($restInfo->match("GET", "authorizations", TRUE)) {
-        $data = $storage->getApproval($restInfo->getResource(), $token->resource_owner_id);
-        $response->setContent(json_encode($data));      
-    } else if($restInfo->match("DELETE", "authorizations", TRUE)) {
-        $data = $storage->deleteApproval($restInfo->getResource(), $token->resource_owner_id);
-    } else if($restInfo->match("GET", "authorizations", FALSE)) {
-        $data = $storage->getApprovals($token->resource_owner_id);
-        $response->setContent(json_encode($data));      
+        if($restInfo->match("POST", "authorizations", FALSE)) {
+            $data = json_decode($request->getContent(), TRUE);
+            $storage->addApproval($data['client_id'], $token->resource_owner_id, $data['scope']);
+            $response->setStatusCode(201);
+        } else if($restInfo->match("GET", "authorizations", TRUE)) {
+            $data = $storage->getApproval($restInfo->getResource(), $token->resource_owner_id);
+            $response->setContent(json_encode($data));      
+        } else if($restInfo->match("DELETE", "authorizations", TRUE)) {
+            $data = $storage->deleteApproval($restInfo->getResource(), $token->resource_owner_id);
+        } else if($restInfo->match("GET", "authorizations", FALSE)) {
+            $data = $storage->getApprovals($token->resource_owner_id);
+            $response->setContent(json_encode($data));      
+        } else {
+            // unsupported call
+            $response->setStatusCode(405);
+            $response->setHeader("Allow", "GET,POST,DELETE");
+            $response->setContent(json_encode(array("error" => "Method Not Allowed")));
+        }
     } else {
-        // unsupported call
-        $response->setContent("ERROR");
-    }
-
+        $response->setStatusCode(401);
+        $response->setHeader("Content-Type", "text/plain");
+        $response->setHeader("WWW-Authenticate", sprintf('Bearer realm="Resource Server"'));
+        $response->setContent("Unauthorized");
+    }   
 } catch (Exception $e) {
     switch(get_class($e)) {
         case "VerifyException":
@@ -50,7 +59,8 @@ try {
             break;
 
         default:
-            echo $e->getMessage();
+            $response->setStatusCode(500);
+            $response->setContent(json_encode(array("error" => $e->getMessage())));
             break;
     }
